third_party/blink/web_tests/external/wpt/css/css-box/margin-trim/computed-margin-values/flexbox-column-block-end-expected.txt

This is a testharness.js-based test.
PASS flexbox > item 1
FAIL flexbox > item 2 assert_equals: 
<item data-expected-margin-bottom="0" data-offset-y="68"></item>
margin-bottom expected "0" but got "10"
Harness: the test ran to completion.
